fix(welllog): make Werror-clean under GCC -Wall -Wextra -Wconversion -Wsign-conversion
- raster_export: cast tile_origin_y int64->double explicitly (3 sites)
- zip_writer: value-initialize DirEntry so struct padding is zeroed
- manifest/renderer: handle triangle_down + shoe SymbolKind switch cases
  (manifest names round-trip as triangleDown/shoe)
- table_projection: isolate document copy in a noinline helper to dodge a
  GCC 16 -O3 -Warray-bounds false positive on the shared_ptr control block
- png_decode.hpp: explicit ptrdiff_t/size_t casts at iterator arithmetic

Verified: Release + Werror build clean, ctest 67/67 pass.

+++ src/export_raster/raster_export.cpp
@@ -285,19 +285,24 @@ void fill_polygon(std::vector<std::uint8_t> &tile, std::uint32_t width,
   }
   const auto x0 = std::max(
       0, static_cast<int>(std::floor((center_x + min_x) * ppm)));
+  // tile_origin_y is a tile offset (uint32 source), so it fits exactly in
+  // double; cast explicitly to keep the mixed int64/double arithmetic honest.
   const auto y0 = std::max(
       0, static_cast<int>(std::floor((center_y + min_y) * ppm) -
-                          tile_origin_y));
+                          static_cast<double>(tile_origin_y)));
   const auto x1 = std::min(
       static_cast<int>(width),
       static_cast<int>(std::ceil((center_x + max_x) * ppm)));
   const auto y1 = std::min(
       static_cast<int>(height),
-      static_cast<int>(std::ceil((center_y + max_y) * ppm) - tile_origin_y));
+      static_cast<int>(std::ceil((center_y + max_y) * ppm) -
+                      static_cast<double>(tile_origin_y)));
   for (int y = y0; y < y1; ++y) {
     for (int x = x0; x < x1; ++x) {
       const auto sx = (static_cast<double>(x) + 0.5) / ppm;
-      const auto sy = (static_cast<double>(y) + 0.5 + tile_origin_y) / ppm;
+      const auto sy = (static_cast<double>(y) + 0.5 +
+                       static_cast<double>(tile_origin_y)) /
+                      ppm;
       if (point_in_polygon(outline, sx - center_x, sy - center_y)) {
         blend_pixel(tile, width, height, channels, x, y, color);
       }
